Refactor project structure (#3)
* Refactor components, resources and events

* Refactor

* Refactoring

* Refactor

* Refactor

* Refactor, compile

* Cargo fix

* Refactor

* Update terrain generator, refactor

* Update imports
