+++ src/collider/components.rs
@@ -1,0 +1,7 @@
+use bevy::ecs::component::Component;
+
+#[derive(Component)]
+pub struct MyCollider {
+    pub key: u32,
+}
+
