+++ src/collider/events.rs
@@ -1,0 +1,6 @@
+use bevy::ecs::event::Event;
+
+#[derive(Event)]
+pub struct ColliderUpdateEvent {
+    pub position: [f32; 3],
+}
