+++ src/player/components.rs
@@ -1,0 +1,4 @@
+use bevy::ecs::component::Component;
+
+#[derive(Component)]
+pub struct HighlightCube;
