+++ src/player/systems/selection.rs
@@ -1,35 +1,28 @@
-use bevy::{
-    asset::Assets,
-    ecs::{
-        component::Component,
-        query::{With, Without},
-        system::{Commands, Query, ResMut, Resource},
-    },
-    gizmos::gizmos::Gizmos,
-    math::{primitives::Cuboid, Ray3d, Vec3},
-    pbr::{PbrBundle, StandardMaterial},
-    prelude::{default, Deref, DerefMut},
-    render::{camera::Camera, color::Color, mesh::Mesh},
-    transform::components::Transform,
-};
+use bevy::{asset::Assets, ecs::{query::{With, Without}, system::{Commands, Query, ResMut}}, gizmos::gizmos::Gizmos, math::{primitives::Cuboid, Ray3d, Vec3}, pbr::{PbrBundle, StandardMaterial}, prelude::default, render::{camera::Camera, color::Color, mesh::Mesh}, transform::components::Transform};
 use bevy_mod_raycast::immediate::{Raycast, RaycastSettings};
 
-#[derive(Resource, Deref, DerefMut)]
-pub struct SelectedPosition(pub Option<Vec3>);
-
-#[derive(Resource)]
-pub struct BlockSelection {
-    pub position: Option<Vec3>,
-    pub normal: Option<Vec3>,
-}
-
-#[derive(Resource, Deref, DerefMut)]
-pub struct SelectedNormal(pub Option<Vec3>);
+use crate::player::{components::HighlightCube, BlockSelection};
 
 const RAY_DIST: Vec3 = Vec3::new(0.0, 0.0, -20.0);
 
-// query camera position and direction
-pub fn raycast(
+pub fn setup_highlight_cube_system(
+    mut commands: Commands,
+    mut meshes: ResMut<Assets<Mesh>>,
+    mut materials: ResMut<Assets<StandardMaterial>>,
+) {
+    let mesh = Cuboid::new(1.01, 1.01, 1.01);
+
+    commands
+        .spawn(PbrBundle {
+            mesh: meshes.add(mesh),
+            material: materials.add(Color::rgba(1.0, 1.0, 1.0, 0.5)),
+            transform: Transform::from_xyz(0.0, 0.0, -7.0),
+            ..default()
+        })
+        .insert(HighlightCube);
+}
+
+pub fn raycast_system(
     mut raycast: Raycast,
     mut gizmos: Gizmos,
     query: Query<&Transform, With<Camera>>,
@@ -70,22 +63,3 @@ pub fn raycast(
     highlight_transform.translation = hover_position.unwrap() + 0.5;
 }
 
-#[derive(Component)]
-pub struct HighlightCube;
-
-pub fn add_highlight_cube(
-    mut commands: Commands,
-    mut meshes: ResMut<Assets<Mesh>>,
-    mut materials: ResMut<Assets<StandardMaterial>>,
-) {
-    let mesh = Cuboid::new(1.01, 1.01, 1.01);
-
-    commands
-        .spawn(PbrBundle {
-            mesh: meshes.add(mesh),
-            material: materials.add(Color::rgba(1.0, 1.0, 1.0, 0.5)),
-            transform: Transform::from_xyz(0.0, 0.0, -7.0),
-            ..default()
-        })
-        .insert(HighlightCube);
-}
