+++ src/terrain/util/blocks.rs
@@ -1,4 +1,4 @@
-use crate::mesher::CubeFace;
+use super::mesher::CubeFace;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum BlockId {
@@ -34,8 +34,7 @@ macro_rules! add_block {
     };
 }
 
-pub static BLOCKS: [Block; 15] = [
-    add_block!(BlockId::Air, ["air"; 6], false),
+pub static BLOCKS: [Block; 14] = [
     add_block!(BlockId::Air, ["air"; 6], false),
     add_block!(
         BlockId::Grass,
